Reorganize `lib` folder
