+++ site/src/lib/course-planner/ClassMeetingUtils.ts
