+++ site/src/lib/course-planner/CourseSearch.ts
